+++ core/upload.go
@@ -18,7 +18,6 @@ import (
 	"crypto/md5"
 	"encoding/json"
 	"fmt"
-	"github.com/gomodule/oauth1/oauth"
 	"io"
 	"io/ioutil"
 	"log"
@@ -29,11 +28,13 @@ import (
 	"path/filepath"
 	"strconv"
 	"time"
+
+	"github.com/gomodule/oauth1/oauth"
 )
 
-const uploadUri = "https://upload.smugmug.com/"
+const uploadURI = "https://upload.smugmug.com/"
 
-type uploadResponseJson struct {
+type uploadResponseJSON struct {
 	Stat    string
 	Message string
 }
@@ -48,7 +49,7 @@ func upload(albumKey string, filename string) {
 
 	var client = http.Client{}
 
-	err = postImage(&client, uploadUri, userToken, albumKey, filename, retriesFlag+1)
+	err = postImage(&client, uploadURI, userToken, albumKey, filename, retriesFlag+1)
 	if err != nil {
 		log.Println("Error uploading: " + err.Error())
 	}
@@ -95,7 +96,7 @@ func multiUpload(numParallel int, albumKey string, filenames []string) {
 		semaph <- 1
 		go func(filename string) {
 			fmt.Println("go " + filename)
-			err := postImage(&client, uploadUri, userToken, albumKey, filename, retriesFlag+1)
+			err := postImage(&client, uploadURI, userToken, albumKey, filename, retriesFlag+1)
 			if err != nil {
 				log.Println("Error uploading: " + err.Error())
 			}
@@ -118,8 +119,8 @@ func getMediaType(filename string) string {
 	return mime.TypeByExtension(ext)
 }
 
-// calcMd5 generates the MD5 sum for the given file.
-func calcMd5(imgFileName string) (string, int64, error) {
+// calcMD5 generates the MD5 sum for the given file.
+func calcMD5(imgFileName string) (string, int64, error) {
 	file, err := os.Open(imgFileName)
 	if err != nil {
 		return "", 0, err
@@ -146,7 +147,7 @@ func postImage(client *http.Client, uri string, credentials *oauth.Credentials,
 	var tryCount uint
 	for tryCount = 0; tryCount < tries; tryCount++ {
 
-		md5Str, imgSize, err := calcMd5(imgFileName)
+		md5Str, imgSize, err := calcMD5(imgFileName)
 		if err != nil {
 			return err
 		}
@@ -211,8 +212,8 @@ func postImage(client *http.Client, uri string, credentials *oauth.Credentials,
 		fmt.Println(resp.Status)
 		fmt.Println(string(bytes))
 
-		var respJson uploadResponseJson
-		err = json.Unmarshal(bytes, &respJson)
+		var respJSON uploadResponseJSON
+		err = json.Unmarshal(bytes, &respJSON)
 		if err != nil {
 			log.Println("Error decoding upload response JSON: " + err.Error())
 			if tryCount < tries-1 {
@@ -221,7 +222,7 @@ func postImage(client *http.Client, uri string, credentials *oauth.Credentials,
 			return err
 		}
 
-		if respJson.Stat == "ok" {
+		if respJSON.Stat == "ok" {
 			break
 		}
 	}